chore: refactor WIP
