+++ packages/pieces/convertkit/README.md
@@ -5,3 +5,14 @@ This library was generated with [Nx](https://nx.dev).
 ## Building
 
 Run `nx build pieces-convertkit` to build the library.
+
+# TODO
+
+- Make sure the proper refreshers are set
+- add remaining webhook triggers
+- add dev switch for webhook using env variable
+- stop flow on errors https://www.activepieces.com/docs/developers/piece-reference/flow-control
+- versioning https://www.activepieces.com/docs/developers/piece-reference/piece-versioning
+- implement retry logic on fetch
+- debounce
+- add tests
