+++ packages/pieces/convertkit/src/lib/actions/broadcasts.ts
@@ -1,73 +1,21 @@
-import { createAction, Property } from '@activepieces/pieces-framework';
+import { createAction } from '@activepieces/pieces-framework';
 import { convertkitAuth } from '../..';
-import { CONVERTKIT_API_URL } from '../common';
-
-const API_ENDPOINT = 'broadcasts';
-
-// const now = new Date();
-
-const broadcastProps = {
-  content: Property.ShortText({
-    displayName: 'Content',
-    description:
-      "The broadcast's email content - this can contain text and simple HTML markdown (such as h1, img or p tags)",
-    required: false,
-  }),
-  description: Property.ShortText({
-    displayName: 'Description',
-    description: 'An internal description of this broadcast',
-    required: false,
-  }),
-  email_address: Property.ShortText({
-    displayName: 'Email Address',
-    description:
-      "Sending email address; leave blank to use your account's default sending email address",
-    required: false,
-  }),
-  email_layout_template: Property.ShortText({
-    displayName: 'Email Layout Template',
-    description:
-      "Name of the email template to use; leave blank to use your account's default email template",
-    required: false,
-  }),
-  public: Property.Checkbox({
-    displayName: 'Public',
-    description: 'Specifies whether or not this is a public post',
-    required: false,
-    defaultValue: false,
-  }),
-  published_at: Property.DateTime({
-    displayName: 'Published At',
-    description:
-      'Specifies the time that this post was published (applicable only to public posts)',
-    required: false,
-    // defaultValue: now.toDateString(),
-  }),
-  send_at: Property.DateTime({
-    displayName: 'Send At',
-    description:
-      'Time that this broadcast should be sent; leave blank to create a draft broadcast. If set to a future time, this is the time that the broadcast will be scheduled to send.',
-    required: false,
-    // defaultValue: now.toDateString(),
-  }),
-  subject: Property.ShortText({
-    displayName: 'Subject',
-    description: "The broadcast email's subject",
-    required: false,
-  }),
-  thumbnail_alt: Property.ShortText({
-    displayName: 'Thumbnail Alt',
-    description:
-      'Specify the ALT attribute of the public thumbnail image (applicable only to public posts)',
-    required: false,
-  }),
-  thumbnail_url: Property.ShortText({
-    displayName: 'Thumbnail Url',
-    description:
-      'Specify the URL of the thumbnail image to accompany the broadcast post (applicable only to public posts)',
-    required: false,
-  }),
-};
+import {
+  API_ENDPOINT,
+  broadcastId,
+  page,
+  content,
+  description,
+  emailAddress,
+  emailLayoutTemplate,
+  isPublic,
+  publishedAt,
+  sendAt,
+  subject,
+  thumbnailAlt,
+  thumbnailUrl,
+} from '../common/broadcasts';
+import { CONVERTKIT_API_URL } from '../common/constants';
 
 export const listBroadcasts = createAction({
   auth: convertkitAuth,
@@ -75,25 +23,26 @@ export const listBroadcasts = createAction({
   displayName: 'Broadcast: List Broadcasts',
   description: 'List all broadcasts',
   props: {
-    page: Property.Number({
-      displayName: 'Page',
-      description:
-        'Page number. Each page of results will contain up to 50 broadcasts.',
-      required: false,
-      defaultValue: 1,
-    }),
+    page,
   },
   async run(context) {
     const page = context.propsValue.page || 1;
-    const url = `${CONVERTKIT_API_URL}${API_ENDPOINT}?page=${page}&api_secret=${context.auth}`;
+    const url = `${CONVERTKIT_API_URL}/${API_ENDPOINT}?page=${page}&api_secret=${context.auth}`;
 
     // Fetch URL using fetch api
     const response = await fetch(url);
 
+    if (!response.ok) {
+      return { success: false, message: 'Error fetching broadcasts' };
+    }
+
     // Get response body
     const data = await response.json();
 
-    // curl command
+    // return broadcasts if exists
+    if (data.broadcasts) {
+      return data.broadcasts;
+    }
 
     // Return response body
     return data;
@@ -105,21 +54,66 @@ export const createBroadcast = createAction({
   name: 'broadcasts_create_broadcast',
   displayName: 'Broadcast: Create Broadcast',
   description: 'Create a new broadcast',
-  props: broadcastProps,
+  props: {
+    content,
+    description,
+    emailAddress,
+    emailLayoutTemplate,
+    isPublic,
+    publishedAt,
+    sendAt,
+    subject,
+    thumbnailAlt,
+    thumbnailUrl,
+  },
   async run(context) {
-    const url = `${CONVERTKIT_API_URL}${API_ENDPOINT}?api_secret=${context.auth}`;
+    const {
+      content,
+      description,
+      emailAddress,
+      emailLayoutTemplate,
+      isPublic,
+      publishedAt,
+      sendAt,
+      subject,
+      thumbnailAlt,
+      thumbnailUrl,
+    } = context.propsValue;
+    const url = `${CONVERTKIT_API_URL}/${API_ENDPOINT}?api_secret=${context.auth}`;
+
+    const body = JSON.stringify({
+      content,
+      description,
+      email_address: emailAddress,
+      email_layout_template: emailLayoutTemplate,
+      public: isPublic,
+      published_at: publishedAt,
+      send_at: sendAt,
+      subject,
+      thumbnail_alt: thumbnailAlt,
+      thumbnail_url: thumbnailUrl,
+    });
 
     // Fetch URL using fetch api
     const response = await fetch(url, {
       method: 'POST',
-      body: JSON.stringify({ ...context.propsValue }),
+      body,
       headers: {
         'Content-Type': 'application/json',
       },
     });
 
+    if (!response.ok) {
+      return { success: false, message: 'Error creating broadcast' };
+    }
+
     // Get response body
     const data = await response.json();
+
+    // if broadcast exists, return it
+    if (data.broadcast) {
+      return data.broadcast;
+    }
 
     // Return response body
     return data;
@@ -132,21 +126,26 @@ export const getBroadcastById = createAction({
   displayName: 'Broadcast: Get Broadcast',
   description: 'Get a broadcast',
   props: {
-    broadcastId: Property.ShortText({
-      displayName: 'Broadcast Id',
-      description: 'The broadcast id',
-      required: true,
-    }),
+    broadcastId,
   },
   async run(context) {
     const { broadcastId } = context.propsValue;
-    const url = `${CONVERTKIT_API_URL}${API_ENDPOINT}/${broadcastId}?api_secret=${context.auth}`;
+    const url = `${CONVERTKIT_API_URL}/${API_ENDPOINT}/${broadcastId}?api_secret=${context.auth}`;
 
     // Fetch URL using fetch api
     const response = await fetch(url);
 
+    if (!response.ok) {
+      return { success: false, message: 'Error fetching broadcast' };
+    }
+
     // Get response body
     const data = await response.json();
+
+    // if broadcast exists, return it
+    if (data.broadcast) {
+      return data.broadcast;
+    }
 
     // Return response body
     return data;
@@ -159,32 +158,69 @@ export const updateBroadcast = createAction({
   displayName: 'Broadcast: Update Broadcast',
   description: 'Update a broadcast',
   props: {
-    broadcastId: Property.ShortText({
-      displayName: 'Broadcast Id',
-      description: 'The broadcast id',
-      required: true,
-    }),
-    ...broadcastProps,
+    broadcastId,
+    content,
+    description,
+    emailAddress,
+    emailLayoutTemplate,
+    isPublic,
+    publishedAt,
+    sendAt,
+    subject,
+    thumbnailAlt,
+    thumbnailUrl,
   },
-  async run(context) {
-    const { broadcastId } = context.propsValue;
-    const url = `${CONVERTKIT_API_URL}${API_ENDPOINT}/${broadcastId}?api_secret=${context.auth}`;
 
-    // TODO: How to remove the 'auth' key?
-    // put all values, except broadcastId, in props
-    const { broadcastId: _, ...props } = context.propsValue;
+  async run(context) {
+    const {
+      broadcastId,
+      content,
+      description,
+      emailAddress,
+      emailLayoutTemplate,
+      isPublic,
+      publishedAt,
+      sendAt,
+      subject,
+      thumbnailAlt,
+      thumbnailUrl,
+    } = context.propsValue;
+
+    const url = `${CONVERTKIT_API_URL}/${API_ENDPOINT}/${broadcastId}?api_secret=${context.auth}`;
+
+    const body = JSON.stringify({
+      content,
+      description,
+      email_address: emailAddress,
+      email_layout_template: emailLayoutTemplate,
+      public: isPublic,
+      published_at: publishedAt,
+      send_at: sendAt,
+      subject,
+      thumbnail_alt: thumbnailAlt,
+      thumbnail_url: thumbnailUrl,
+    });
 
     // Fetch URL using fetch api
     const response = await fetch(url, {
       method: 'PUT',
-      body: JSON.stringify({ ...props }),
+      body,
       headers: {
         'Content-Type': 'application/json',
       },
     });
 
+    if (!response.ok) {
+      return { success: false, message: 'Error updating broadcast' };
+    }
+
     // Get response body
     const data = await response.json();
+
+    // if broadcast exists, return it
+    if (data.broadcast) {
+      return data.broadcast;
+    }
 
     // Return response body
     return data;
@@ -197,21 +233,26 @@ export const broadcastStats = createAction({
   displayName: 'Broadcast: Broadcast Stats',
   description: 'Get broadcast stats',
   props: {
-    broadcastId: Property.ShortText({
-      displayName: 'Broadcast Id',
-      description: 'The broadcast id',
-      required: true,
-    }),
+    broadcastId,
   },
   async run(context) {
     const { broadcastId } = context.propsValue;
-    const url = `${CONVERTKIT_API_URL}${API_ENDPOINT}/${broadcastId}/stats?api_secret=${context.auth}`;
+    const url = `${CONVERTKIT_API_URL}/${API_ENDPOINT}/${broadcastId}/stats?api_secret=${context.auth}`;
 
     // Fetch URL using fetch api
     const response = await fetch(url);
 
+    if (!response.ok) {
+      return { success: false, message: 'Error fetching broadcast stats' };
+    }
+
     // Get response body
     const data = await response.json();
+
+    // if broadcast exists, return it
+    if (data.broadcast) {
+      return data.broadcast;
+    }
 
     // Return response body
     return data;
@@ -224,15 +265,11 @@ export const deleteBroadcast = createAction({
   displayName: 'Broadcast: Delete Broadcast',
   description: 'Delete a broadcast',
   props: {
-    broadcastId: Property.ShortText({
-      displayName: 'Broadcast Id',
-      description: 'The broadcast id',
-      required: true,
-    }),
+    broadcastId,
   },
   async run(context) {
     const { broadcastId } = context.propsValue;
-    const url = `${CONVERTKIT_API_URL}${API_ENDPOINT}/${broadcastId}?api_secret=${context.auth}`;
+    const url = `${CONVERTKIT_API_URL}/${API_ENDPOINT}/${broadcastId}?api_secret=${context.auth}`;
 
     // Fetch URL using fetch api
     const response = await fetch(url, {
