+++ packages/pieces/convertkit/src/lib/actions/custom-fields.ts
@@ -1,18 +1,13 @@
-import {
-  createAction,
-  Property,
-  DynamicPropsValue,
-} from '@activepieces/pieces-framework';
+import { createAction } from '@activepieces/pieces-framework';
 import { convertkitAuth } from '../..';
-import { CONVERTKIT_API_URL } from '../common';
-
-const API_ENDPOINT = 'custom_fields';
-
-export const getCustomFields = async (auth: string) => {
-  const url = `${CONVERTKIT_API_URL}${API_ENDPOINT}?api_secret=${auth}`;
-  const response = await fetch(url);
-  return await response.json();
-}
+import {
+  API_ENDPOINT,
+  fetchCustomFields,
+  fieldsArray,
+  label,
+  newLabel,
+} from '../common/custom-fields';
+import { CONVERTKIT_API_URL } from '../common/constants';
 
 export const listFields = createAction({
   auth: convertkitAuth,
@@ -21,42 +16,12 @@ export const listFields = createAction({
   description: 'Returns a list of all custom fields',
   props: {},
   async run(context) {
-    return getCustomFields(context.auth);
-  },
-});
-
-
-export const propertyCustomFields =  Property.DynamicProperties({
-  displayName: 'Custom Fields',
-  description: 'The custom fields',
-  required: false,
-  refreshers: ['auth'],
-  props: async ({ auth }) => {
-    if (!auth) {
-      return {
-        disabled: true,
-        placeholder: 'Connect your account',
-        options: [],
-      };
+    const data = await fetchCustomFields(context.auth);
+    // if custom_fields exist, return custom_fields
+    if (data.custom_fields) {
+      return data.custom_fields;
     }
-
-    const fields: DynamicPropsValue = {};
-
-    const customFields = await getCustomFields(auth.toString());
-
-    // loop through data and map to fields
-    customFields.custom_fields.forEach(
-      (field: { id: string; label: string; key: string; name: string }) => {
-        fields[field.key] = Property.ShortText({
-          displayName: field.label,
-          description: `Enter the value for custom field: ${field.label}`,
-          required: false,
-        });
-      }
-    );
-
-    console.debug('fields: ', fields);
-    return fields;
+    return data;
   },
 });
 
@@ -66,15 +31,10 @@ export const createField = createAction({
   displayName: 'Custom Fields: Create Field',
   description: 'Create a new custom field',
   props: {
-    // TODO: Add validation for the fields
-    fields: Property.Array({
-      displayName: 'Fields',
-      description: 'The custom fields',
-      required: true,
-    }),
+    fields: fieldsArray,
   },
   async run(context) {
-    const url = `${CONVERTKIT_API_URL}${API_ENDPOINT}?api_secret=${context.auth}`;
+    const url = `${CONVERTKIT_API_URL}/${API_ENDPOINT}?api_secret=${context.auth}`;
 
     // Fetch URL using fetch api
     const response = await fetch(url, {
@@ -84,6 +44,10 @@ export const createField = createAction({
         'Content-Type': 'application/json',
       },
     });
+
+    if (!response.ok) {
+      return { success: false, message: 'Error creating field' };
+    }
 
     // Get response body
     const data = await response.json();
@@ -99,67 +63,13 @@ export const updateField = createAction({
   displayName: 'Custom Fields: Update Field',
   description: 'Update a custom field',
   props: {
-    label: Property.Dropdown({
-      displayName: 'Custom Label',
-      required: true,
-      refreshers: ['auth', 'new_label'],
-      options: async ({ auth }) => {
-        if (!auth) {
-          return {
-            disabled: true,
-            placeholder: 'Connect your account',
-            options: [],
-          };
-        }
-
-        const fields = await getCustomFields(auth.toString());
-
-        // loop through data and map to options
-        const options = fields.custom_fields.map(
-          (field: { id: string; label: string; key: string; name: string }) => {
-            return {
-              label: field.label,
-              value: field.id,
-            };
-          }
-        );
-
-        return {
-          options,
-        };
-      },
-    }),
-    new_label: Property.DynamicProperties({
-      // TODO: refresh after lables are updated. is it possible/useful?
-      displayName: 'Event Parameter',
-      description: 'The required parameter for the event',
-      required: true,
-      refreshers: ['label', 'auth'],
-      props: async ({ auth, label }) => {
-        if (!auth) {
-          return {
-            disabled: true,
-            placeholder: 'Select labels first',
-            options: [],
-          };
-        }
-
-        const fields: DynamicPropsValue = {};
-
-        fields[label.toString()] = Property.ShortText({
-          displayName: `New label`,
-          // description: `Enter the new label for ${customField.label}`,
-          required: true,
-        });
-        console.debug('fields: ', fields);
-        return fields;
-      },
-    }),
+    label,
+    new_label: newLabel,
   },
   async run(context) {
     const { label, new_label } = context.propsValue;
 
-    const url = `${CONVERTKIT_API_URL}${API_ENDPOINT}/${label}`;
+    const url = `${CONVERTKIT_API_URL}/${API_ENDPOINT}/${label}`;
 
     // Fetch URL using fetch api
     const response = await fetch(url, {
@@ -186,43 +96,12 @@ export const deleteField = createAction({
   displayName: 'Custom Fields: Delete Field',
   description: 'Delete a custom field',
   props: {
-    label: Property.Dropdown({
-      displayName: 'Custom Label',
-      required: true,
-      refreshers: ['auth'],
-      options: async ({ auth }) => {
-        if (!auth) {
-          return {
-            disabled: true,
-            placeholder: 'Connect your account',
-            options: [],
-          };
-        }
-
-        const url = `${CONVERTKIT_API_URL}${API_ENDPOINT}?api_secret=${auth}`;
-        const response = await fetch(url);
-        const data = await response.json();
-
-        // loop through data and map to options
-        const options = data.custom_fields.map(
-          (field: { id: string; label: string; key: string; name: string }) => {
-            return {
-              label: field.label,
-              value: field.id,
-            };
-          }
-        );
-
-        return {
-          options,
-        };
-      },
-    }),
+    label,
   },
   async run(context) {
     const { label } = context.propsValue;
 
-    const url = `${CONVERTKIT_API_URL}${API_ENDPOINT}/${label}`;
+    const url = `${CONVERTKIT_API_URL}/${API_ENDPOINT}/${label}`;
 
     // Fetch URL using fetch api
     const response = await fetch(url, {
