+++ packages/pieces/convertkit/src/lib/actions/sequences.ts
@@ -1,47 +1,15 @@
-import { createAction, Property } from '@activepieces/pieces-framework';
+import { createAction } from '@activepieces/pieces-framework';
 import { convertkitAuth } from '../..';
-import { CONVERTKIT_API_URL } from '../common';
-import { propertyCustomFields } from './custom-fields';
-import { propertyTags } from './tags';
-
-const API_ENDPOINT = 'sequences';
-
-export const getSequences = async (auth: string) => {
-  const url = `${CONVERTKIT_API_URL}${API_ENDPOINT}?api_secret=${auth}`;
-  const response = await fetch(url);
-  return await response.json();
-};
-
-export const propertySequence = Property.Dropdown({
-  displayName: 'Sequence',
-  required: true,
-  refreshers: ['auth'],
-  options: async ({ auth }) => {
-    if (!auth) {
-      return {
-        disabled: true,
-        placeholder: 'Connect your account',
-        options: [],
-      };
-    }
-
-    const sequences = await getSequences(auth.toString());
-
-    // loop through data and map to options
-    const options = sequences.courses.map(
-      (field: { id: string; name: string }) => {
-        return {
-          label: field.name,
-          value: field.id,
-        };
-      }
-    );
-
-    return {
-      options,
-    };
-  },
-});
+import {
+  API_ENDPOINT,
+  fetchSequences,
+  sequenceIdChoice,
+  firstName,
+  email,
+} from '../common/sequences';
+import { allFields } from '../common/custom-fields';
+import { tags } from '../common//tags';
+import { CONVERTKIT_API_URL } from '../common/constants';
 
 export const listSequences = createAction({
   auth: convertkitAuth,
@@ -50,7 +18,12 @@ export const listSequences = createAction({
   description: 'Returns a list of all sequences',
   props: {},
   async run(context) {
-    return getSequences(context.auth);
+    const data = await fetchSequences(context.auth);
+    // if courses exist, return courses
+    if (data.courses) {
+      return data.courses;
+    }
+    return data;
   },
 });
 
@@ -60,31 +33,30 @@ export const addSubscriberToSequence = createAction({
   displayName: 'Sequences: Add Subscriber To Sequence',
   description: 'Add a subscriber to a sequence',
   props: {
-    sequenceId: propertySequence,
-    email: Property.ShortText({
-      displayName: 'Email',
-      description: 'The email of the subscriber',
-      required: true,
-    }),
-    first_name: Property.ShortText({
-      displayName: 'First Name',
-      description: 'The first name of the subscriber',
-      required: false,
-    }),
-    tags: propertyTags,
-    fields: propertyCustomFields,
+    sequenceId: sequenceIdChoice,
+    email,
+    firstName,
+    tags,
+    fields: allFields,
   },
   async run(context) {
-    const url = `${CONVERTKIT_API_URL}${API_ENDPOINT}/${context.propsValue.sequenceId}/subscribe?api_secret=${context.auth}`;
+    const { sequenceId, email, firstName, tags, fields } = context.propsValue;
+    const url = `${CONVERTKIT_API_URL}/${API_ENDPOINT}/${context.propsValue.sequenceId}/subscribe?api_secret=${context.auth}`;
+
+    const body = JSON.stringify({ email, first_name: firstName, tags, fields });
 
     // Fetch URL using fetch api
     const response = await fetch(url, {
       method: 'POST',
-      body: JSON.stringify({ ...context.propsValue }),
+      body,
       headers: {
         'Content-Type': 'application/json',
       },
     });
+
+    if (!response.ok) {
+      return { success: false, message: 'Error adding subscriber to sequence' };
+    }
 
     // Get response body
     const data = await response.json();
@@ -100,13 +72,20 @@ export const listSupscriptionsToSequence = createAction({
   displayName: 'Sequences: List Subscriptions To Sequence',
   description: 'List all subscriptions to a sequence',
   props: {
-    sequenceId: propertySequence,
+    sequenceId: sequenceIdChoice,
   },
   async run(context) {
-    const url = `${CONVERTKIT_API_URL}${API_ENDPOINT}/${context.propsValue.sequenceId}/subscriptions?api_secret=${context.auth}`;
+    const url = `${CONVERTKIT_API_URL}/${API_ENDPOINT}/${context.propsValue.sequenceId}/subscriptions?api_secret=${context.auth}`;
 
     // Fetch URL using fetch api
     const response = await fetch(url);
+
+    if (!response.ok) {
+      return {
+        success: false,
+        message: 'Error listing subscriptions to sequence',
+      };
+    }
 
     // Get response body
     const data = await response.json();
