+++ packages/pieces/convertkit/src/lib/common/constants.ts
@@ -1,0 +1,1 @@
+export const CONVERTKIT_API_URL = 'https://api.convertkit.com/v3';
