+++ packages/pieces/convertkit/src/lib/common/index.ts
@@ -1,9 +1,8 @@
-import { Property } from '@activepieces/pieces-framework';
-
-export const CONVERTKIT_API_URL = 'https://api.convertkit.com/v3/';
-
-export const subscriberId = Property.Number({
-  displayName: 'Subscriber ID',
-  description: 'The subscriber ID',
-  required: true,
-});
+export * as broadcasts from './broadcasts';
+export * as customFields from './custom-fields';
+export * as forms from './forms';
+export * as purchases from './purchases';
+export * as subscribers from './subscribers';
+export * as tags from './tags';
+export * as sequences from './sequences';
+export * as webhooks from './webhooks';
